refactor: add ios extensions
