+++ PrePurchaseSDK+Extension.swift
@@ -1,0 +1,34 @@
+import TicketmasterFoundation
+import TicketmasterAuthentication
+import TicketmasterDiscoveryAPI
+import TicketmasterPrePurchase
+import TicketmasterPurchase
+
+extension PrePurchaseSDK: TMPrePurchaseNavigationDelegate, TMPrePurchaseAnalyticsDelegate {
+  
+  public func prePurchaseViewController(_ viewController: TicketmasterPrePurchase.TMPrePurchaseViewController, didShare pageTitle: String, and pageURL: URL, to activityType: UIActivity.ActivityType) {
+    return
+  }
+  
+  public func prePurchaseViewController(_ viewController: TicketmasterPrePurchase.TMPrePurchaseViewController, didFirePageView pageView: TicketmasterFoundation.UALPageView) {
+    //    sendEvent("igniteAnalytics", body: ["prePurchaseSdkDidFirePageView": "\(pageView)"])
+    return
+  }
+  
+  public func prePurchaseViewController(_ viewController: TicketmasterPrePurchase.TMPrePurchaseViewController, navigateToEventDetailsPageWithIdentifier eventIdentifier: String) {
+    let configuration = TMPurchaseWebsiteConfiguration(eventID: eventIdentifier)
+    let apiKey = Config.shared.get(for: "apiKey")
+    let region = Config.shared.get(for: "region")
+    let primaryColor = Config.shared.get(for: "primaryColor")
+    let backgroundColor = UIColor(hexString: primaryColor) ?? AppConstants.defaultBrandColor
+    TMPurchase.shared.configure(apiKey: apiKey, region: TMAuthentication.TMXDeploymentRegion(rawValue: region) ?? .US, completion: {isPurchaseApiSet in
+      TMPurchase.shared.brandColor = backgroundColor!
+      TMDiscoveryAPI.shared.configure(apiKey: apiKey, region: TMAuthentication.TMXDeploymentRegion(rawValue: region) ?? .US, completion: { isDiscoveryApiSet in
+        if (isDiscoveryApiSet && isPurchaseApiSet) {
+          let purchaseNavController = TMPurchaseNavigationController(configuration: configuration)
+          viewController.present(purchaseNavController, animated: true, completion: nil)
+        }
+      })
+    })
+  }
+}
